Implementado Registro de Mascotas: -implementados controladores para registrar Mascota nueva -arreglada e implementada la vista inscribirMascota.html -arreglado un bug al crear usuario por el tipo de id de Role.java

+++ src/main/java/com/unal/una_huella/UNaHuellaLauncher/Entities/Role.java
@@ -12,7 +12,7 @@ public class Role implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO, generator = "native")
     @GenericGenerator(name = "native", strategy = "native")
-    private int id;
+    private Long id;
 
     @Column
     private String name;
@@ -20,11 +20,11 @@ public class Role implements Serializable {
     @Column
     private String description;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id_role) {
+    public void setId(Long id_role) {
         this.id = id_role;
     }
